GPX import worked!
The GPX import has finnaly been made to work. This has been done by changing the gatherGeneral.py file to output time as a timestamp and not a duration.

This can then be more esily translated by the FeatureSet class into the same format as the files converted from the database.